Refactor cards to a separate directory with common props
