+++ src/cards/SpeciesCard.js
@@ -2,6 +2,8 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import EbiSpeciesIcon from 'react-ebi-species'
 
+import cardPropTypes from './CardPropTypes'
+
 const SpeciesCard = ({iconSrc, description, content}) => {
   return (
     <div className={`column column-block text-center combo card`} style={{marginBottom:0, paddingBottom: `25px`}}>
@@ -41,16 +43,6 @@ const SpeciesCard = ({iconSrc, description, content}) => {
   )
 }
 
-SpeciesCard.propTypes = {
-  iconSrc: PropTypes.string.isRequired,
-  description: PropTypes.shape({
-    text: PropTypes.string.isRequired,
-    url: PropTypes.string
-  }),
-  content: PropTypes.arrayOf(PropTypes.shape({
-    text: PropTypes.string.isRequired,
-    url: PropTypes.string
-  }))
-}
+SpeciesCard.propTypes = cardPropTypes
 
 export default SpeciesCard
